create getTag

+++ backend/src/main/java/com/softserveinc/ita/homeprojectblog/repository/TagRepository.java
@@ -4,11 +4,10 @@ import com.softserveinc.ita.homeprojectblog.entity.TagEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigDecimal;
 import java.util.Optional;
 
 @Repository
-public interface TagRepository extends JpaRepository<TagEntity, BigDecimal> {
+public interface TagRepository extends JpaRepository<TagEntity, Long> {
 
     Optional<TagEntity> findByName(String name);
 }
